central_stores/requirements.txt
# Core
Django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.6  # For PostgreSQL (optional)
dj-database-url==2.0.0  # For DB configuration

# Authentication
djangorestframework-simplejwt==5.2.2  # JWT support
dj-rest-auth==3.0.0

# Blockchain
web3==6.0.0
eth-account==0.9.0
pycoin==0.92.20230428

# API Documentation
drf-spectacular==0.26.2  # Swagger/OpenAPI docs
drf-spectacular-sidecar==2023.9.1  # Static files for docs

# Development
django-debug-toolbar==3.8.1
django-cors-headers==4.2.0  # For frontend connections
pylint-django==2.5.3

# Testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.2.1

# Async (optional)
channels==4.0.0  # For WebSocket support
channels-redis==4.1.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0
django-environ==0.11.2